requirements.txt
PyPDF2
streamlit==1.34
python-dotenv
openai
pyautogen
st-gsheets-connection
google-generativeai==0.3.1
pdfminer.six
streamlit-pdf-viewer==0.0.12
firebase_admin==0.14.0
pdfplumber==0.11.0
deepeval==0.21.48